pyproject.toml
[project]
name = "conta-bot-server"
version = "0.1.0"
description = "Telegram bot + AI agent for expense tracking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "python-telegram-bot>=21.0",
    "claude-agent-sdk",
    "pocketbase>=0.13.0",
    "groq>=0.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
